seperate mock from test file to util file
