+++ plan/BUILD.bazel
@@ -38,6 +38,7 @@ go_library(
         "//ast:go_default_library",
         "//context:go_default_library",
         "//distsql:go_default_library",
+        "//domain:go_default_library",
         "//expression:go_default_library",
         "//infoschema:go_default_library",
         "//kv:go_default_library",
@@ -50,16 +51,19 @@ go_library(
         "//sessionctx/variable:go_default_library",
         "//sessionctx/varsutil:go_default_library",
         "//statistics:go_default_library",
+        "//store/tikv/oracle:go_default_library",
         "//table:go_default_library",
         "//tablecodec:go_default_library",
         "//terror:go_default_library",
         "//util/charset:go_default_library",
         "//util/codec:go_default_library",
+        "//util/mock:go_default_library",
         "//util/ranger:go_default_library",
         "//util/types:go_default_library",
         "//vendor/github.com/juju/errors:go_default_library",
         "//vendor/github.com/ngaut/log:go_default_library",
         "//vendor/github.com/pingcap/tipb/go-tipb:go_default_library",
+        "//vendor/golang.org/x/net/context:go_default_library",
     ],
 )
 
@@ -75,17 +79,14 @@ go_test(
     deps = [
         "//ast:go_default_library",
         "//context:go_default_library",
-        "//domain:go_default_library",
         "//expression:go_default_library",
         "//infoschema:go_default_library",
-        "//kv:go_default_library",
         "//model:go_default_library",
         "//mysql:go_default_library",
         "//parser:go_default_library",
         "//sessionctx:go_default_library",
         "//sessionctx/variable:go_default_library",
         "//statistics:go_default_library",
-        "//store/tikv/oracle:go_default_library",
         "//terror:go_default_library",
         "//util/charset:go_default_library",
         "//util/mock:go_default_library",
@@ -94,8 +95,6 @@ go_test(
         "//util/testutil:go_default_library",
         "//util/types:go_default_library",
         "//vendor/github.com/pingcap/check:go_default_library",
-        "//vendor/github.com/pingcap/tipb/go-tipb:go_default_library",
-        "//vendor/golang.org/x/net/context:go_default_library",
     ],
 )
